+++ lib/src/Benchmark/benchmark.dart
@@ -6,7 +6,7 @@ abstract class Benchmark {
   
   void setUp(BenchFunc func);
   
-  void bench(BenchFunc func, [String name, String unit = MILLISECONDS]);
+  void bench(BenchFunc func, [String name, int interations, String unit = MILLISECONDS]);
   
   void tearDown(BenchFunc func);
   
@@ -43,8 +43,9 @@ class _BenchInfo {
   final BenchFunc bench;
   final String name;
   final String unit;
+  final int interations;
   
-  const _BenchInfo(this.bench, this.name, this.unit);
+  const _BenchInfo(this.bench, this.name, this.interations, this.unit);
 }
 
 class _BenchmarkImpl implements Benchmark {
@@ -63,10 +64,11 @@ class _BenchmarkImpl implements Benchmark {
     this._setUp = func;
   }
   
-  void bench(BenchFunc func, [String name, String unit = MILLISECONDS]) {
+  void bench(BenchFunc func, [String name, int interations, String unit = MILLISECONDS]) {
     if(name == null) name = "${this._benchInfos.length + 1}";
+    if(interations == null) interations = this.interations;
     
-    this._benchInfos.add(new _BenchInfo(func, name, unit));
+    this._benchInfos.add(new _BenchInfo(func, name, interations, unit));
   }
   
   void tearDown(BenchFunc func) {
@@ -75,6 +77,8 @@ class _BenchmarkImpl implements Benchmark {
   
   Future run() {
     
+    BenchContext.output.startBenchmark(this.name);
+    
     if(this._benchInfos.length == 0) return null;
     var context = new _BenchContextImpl();
     
@@ -82,7 +86,7 @@ class _BenchmarkImpl implements Benchmark {
         if(this._setUp != null) return this._setUp(context); 
       }).then((_) => Future.forEach(_benchInfos, (benchinfo) {
       
-      var list = new List<_BenchInfo>.filled(this.interations, benchinfo);
+      var list = new List<_BenchInfo>.filled(benchinfo.interations, benchinfo);
       var benchResult = new BenchResult(benchinfo.unit);
       
       return Future.forEach(list, (bench) { 
@@ -90,11 +94,11 @@ class _BenchmarkImpl implements Benchmark {
         return measure((_) => bench.bench(context), bench.unit).then((time) => 
             benchResult.add(time));
         
-      }).then((_) => 
-          print("${this.name} -  ${benchinfo.name} - $benchResult"));
+      }).then((_) => BenchContext.output.printBenchmarkResult(benchinfo.name, benchResult));
       
     })).whenComplete(() {
       if(this._tearDown != null) return this._tearDown(context); 
+      BenchContext.output.endBenchmark();
     });
     
   }